+++ src/scripts/components/legacy-stories/youtube-player/youtube-player.tsx
@@ -1,8 +1,10 @@
 import { FunctionComponent } from "react";
+import { FormattedMessage } from "react-intl";
 import YouTube, { Options } from "react-youtube";
 import { YouTubePlayer } from "youtube-player/dist/types";
-
 import { Language } from "../../../types/language";
+
+import { loadConsent } from "../../../libs/load-consent";
 
 import styles from "./youtube-player.module.css";
 
@@ -19,6 +21,14 @@ const YoutubePlayer: FunctionComponent<Props> = ({
   isStoryMode,
   onPlay,
 }) => {
+  if (!loadConsent()?.youTube) {
+    return (
+      <div className={styles.consentMessage}>
+        <FormattedMessage id="youTube.consentRequired" />
+      </div>
+    );
+  }
+
   const options: Options = {
     height: "100%",
     width: "100%",