+++ src/scripts/types/app-routes.ts
@@ -2,6 +2,7 @@ export enum AppRoute {
   About = "about",
   Base = "base",
   Data = "data",
+  Search = "search",
   Present = "present",
   PresentStory = "present_story",
   Showcase = "showcase",