+++ src/scripts/types/cookie-consent.ts
@@ -1,0 +1,5 @@
+export interface CookieConsent {
+  necessary: boolean;
+  analytics: boolean;
+  youTube: boolean;
+}